[Bugfix][LoRA][Operator] Patch for lora register_model issue after overriding VocabParallelEmbedding class (#2515)
